Extract modules to a separate file
